reformat authentication process
