+++ client/src/components/ModalGraphName/ModalGraphName.jsx
@@ -1,0 +1,5 @@
+import React, { userState, useEffect} from 'react';
+
+// function ModalGraphName = () => {
+
+// }
